update dataset & preprocessor builders, fix compression
This is a major commit that applies various updates to the
project. Major updates include moving the dataset builder
to the tf.data API. Some variable and name scoping is
also improved. Finally, resolves #14 by fixing
the compression library. Updates docs to reflect the need
to export before running compression.